+++ kittens/diff/collect.py
@@ -7,7 +7,7 @@ import re
 from contextlib import suppress
 from functools import lru_cache
 from hashlib import md5
-from mimetypes import guess_type
+from kitty.guess_mime_type import guess_type
 from typing import TYPE_CHECKING, Dict, List, Set, Optional, Iterator, Tuple, Union
 
 if TYPE_CHECKING:
@@ -136,7 +136,7 @@ def sanitize(text: str) -> str:
 
 @lru_cache(maxsize=1024)
 def mime_type_for_path(path: str) -> str:
-    return guess_type(path)[0] or 'application/octet-stream'
+    return guess_type(path) or 'application/octet-stream'
 
 
 @lru_cache(maxsize=1024)