+++ kitty/complete.py
@@ -394,10 +394,10 @@ def complete_files_and_dirs(
 
 
 def complete_icat_args(ans: Completions, opt: Optional[OptionDict], prefix: str, unknown_args: Delegate) -> None:
-    from mimetypes import guess_type
+    from .guess_mime_type import guess_type
 
     def icat_file_predicate(filename: str) -> bool:
-        mt = guess_type(filename)[0]
+        mt = guess_type(filename)
         if mt and mt.startswith('image/'):
             return True
         return False