+++ docs/changelog.rst
@@ -24,6 +24,9 @@ To update |kitty|, :doc:`follow the instructions <binary>`.
 - Fix selections created by dragging upwards not being auto-cleared when
   screen contents change (:pull:`3028`)
 
+- Allow adding MIME definitions to kitty by placing a ``mime.types`` file in
+  the kitty config directory (:iss:`3056`)
+
 
 0.19.1 [2020-10-06]
 -------------------